# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studio Layse Duarte API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get studio settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update studio settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /schedule:
    get:
      operationId: getSchedule
      tags: [schedule]
      summary: Get working hours schedule
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHours"
    put:
      operationId: updateSchedule
      tags: [schedule]
      summary: Update working hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UpdateWorkingHoursInput"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkingHours"
  /availability:
    get:
      operationId: getAvailability
      tags: [schedule]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /public/book:
    post:
      operationId: publicBook
      tags: [appointments]
      summary: Public booking endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationMinutes:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - durationMinutes
        - createdAt
    CreateServiceInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        durationMinutes:
          type: integer
      required:
        - name
        - price
        - durationMinutes
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        serviceId:
          type: integer
        serviceName:
          type: string
        servicePrice:
          type: number
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [confirmed, cancelled, completed, blocked]
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
          enum: [pix, cash, card, null]
        createdAt:
          type: string
      required:
        - id
        - clientName
        - serviceId
        - serviceName
        - servicePrice
        - date
        - startTime
        - endTime
        - status
        - createdAt
...
    Settings:
      type: object
      properties:
        studioName:
          type: string
        primaryColor:
          type: string
        welcomeMessage:
          type: string
          nullable: true
        bookingMessage:
          type: string
          nullable: true
      required:
        - studioName
        - primaryColor
    UpdateSettingsInput:
      type: object
      properties:
        studioName:
          type: string
        primaryColor:
          type: string
        welcomeMessage:
          type: string
          nullable: true
        bookingMessage:
          type: string
          nullable: true
    WorkingHours:
      type: object
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
        dayName:
          type: string
        isOpen:
          type: boolean
        openTime:
          type: string
          nullable: true
        closeTime:
          type: string
          nullable: true
        breakStart:
          type: string
          nullable: true
        breakEnd:
          type: string
          nullable: true
      required:
        - id
        - dayOfWeek
        - dayName
        - isOpen
    UpdateWorkingHoursInput:
      type: object
      properties:
        dayOfWeek:
          type: integer
        isOpen:
          type: boolean
        openTime:
          type: string
          nullable: true
        closeTime:
          type: string
          nullable: true
        breakStart:
          type: string
          nullable: true
        breakEnd:
          type: string
          nullable: true
      required:
        - dayOfWeek
        - isOpen
    PublicBookingInput:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        clientEmail:
          type: string
          nullable: true
        serviceId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - clientName
        - serviceId
        - date
        - startTime