# lib/api-zod/src/generated/types/publicBookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicBookingInput {
  clientName: string;
  clientPhone?: string | null;
  clientEmail?: string | null;
  serviceId: number;
  date: string;
  startTime: string;
  notes?: string | null;
}