# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  durationMinutes: number;
  createdAt: string;
}