# lib/api-zod/src/generated/types/workingHours.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkingHours {
  id: number;
  dayOfWeek: number;
  dayName: string;
  isOpen: boolean;
  openTime?: string | null;
  closeTime?: string | null;
  breakStart?: string | null;
  breakEnd?: string | null;
}